Changed Room.contents to be an Inventory object
Room contents will behave very similarly to an inventory,
so I changed it from an array to an Inventory object so
I have all of the related methods that interact with Item
objects. May have to change it later to be a subclass
of Inventory so that messages can be easily changed based
on whether it's the player's inv or a room's contents.